math: fix implicit conversion
In file included from ../include/text.h:4:
../include/math.h:122:27: error: implicit conversion changes signedness: '__size_t' (aka 'unsigned long') to 's32' (aka 'int') [-Werror,-Wsign-conversion]
  122 |   return __builtin_clz(x) ^ (sizeof(s32) * 8 - 1);
      |   ~~~~~~ ~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~
../include/math.h:122:10: error: implicit conversion changes signedness: 'int' to '__size_t' (aka 'unsigned long') [-Werror,-Wsign-conversion]
  122 |   return __builtin_clz(x) ^ (sizeof(s32) * 8 - 1);
      |          ^~~~~~~~~~~~~~~~ ~
../include/math.h:127:10: error: implicit conversion changes signedness: 'int' to '__size_t' (aka 'unsigned long') [-Werror,-Wsign-conversion]
  127 |   return __builtin_clzll(x) ^ (sizeof(s64) * 8 - 1);
      |          ^~~~~~~~~~~~~~~~~~ ~
3 errors generated.

+++ include/math.h
@@ -119,12 +119,12 @@ Lerp(f32 a, f32 b, f32 t)
 static inline s32
 bsr(u32 x)
 {
-  return __builtin_clz(x) ^ (sizeof(s32) * 8 - 1);
+  return __builtin_clz(x) ^ (int)(sizeof(s32) * 8 - 1);
 }
 static inline s64
 bsrl(u64 x)
 {
-  return __builtin_clzll(x) ^ (sizeof(s64) * 8 - 1);
+  return __builtin_clzll(x) ^ (int)(sizeof(s64) * 8 - 1);
 }
 
 static inline f32
